Move Otp::generate() into new class OtpGenerator.
