+++ src/Surfnet/StepupSelfService/SelfServiceBundle/Tests/Service/SmsSecondFactor/OtpGeneratorTest.php
@@ -1,0 +1,35 @@
+<?php
+
+/**
+ * Copyright 2014 SURFnet bv
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+namespace Surfnet\StepupSelfService\SelfServiceBundle\Tests\Service\SmsSecondFactor;
+
+use PHPUnit_Framework_TestCase as TestCase;
+use Surfnet\StepupSelfService\SelfServiceBundle\Service\SmsSecondFactor\OtpGenerator;
+
+final class OtpGeneratorTest extends TestCase
+{
+    /**
+     * @test
+     * @group sms
+     */
+    public function it_generates_otp_strings()
+    {
+        $this->assertNotEmpty(OtpGenerator::generate());
+        $this->assertInternalType('string', OtpGenerator::generate());
+    }
+}
